+++ .gitignore
@@ -3,3 +3,6 @@
 *.pyo
 __pycache__
 venv
+db.sqlite3
+test_case_files
+
