+++ backend/drf_backend/tcms/models/test_case.py
@@ -89,23 +89,6 @@ class BugTrackerTickets(models.Model):
 
 
 
-class TestSuite(models.Model):
-    test_suite_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255, unique=True)
-    creator_id = models.IntegerField()
-    description = models.TextField(null=True, blank=True)
-
-class Section(models.Model):
-    section_id = models.AutoField(primary_key=True)
-    test_suite = models.ForeignKey(TestSuite, on_delete=models.CASCADE)
-    name = models.CharField(max_length=255, unique=True)
-    creator_id = models.IntegerField()
-    subsection_id = models.IntegerField(null=True, blank=True)
-
-class TestSuiteFiles(models.Model):
-    file_id = models.AutoField(primary_key=True)
-    file_path = models.CharField(max_length=255)
-    test_suite = models.ForeignKey(TestSuite, on_delete=models.CASCADE)
 
 
 
@@ -113,73 +96,6 @@ class TestSuiteFiles(models.Model):
 
 
 
-class TestRun(models.Model):
-    test_run_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255, unique=True)
-    test_suite = models.ForeignKey(TestSuite, on_delete=models.CASCADE)
-    creator_id = models.IntegerField()
-    milestone_id = models.IntegerField(null=True, blank=True)
-    description = models.TextField(null=True, blank=True)
-    test_case_filter = models.TextField(null=True, blank=True)
-
-class TestRunFiles(models.Model):
-    file_id = models.AutoField(primary_key=True)
-    file_path = models.CharField(max_length=255)
-    test_run = models.ForeignKey(TestRun, on_delete=models.CASCADE)
-
-class TestRunTicket(models.Model):
-    ticket_id = models.AutoField(primary_key=True)
-    ticket = models.CharField(max_length=255, unique=True)
-    test_run = models.ForeignKey(TestRun, on_delete=models.CASCADE)
-
-class TestPlan(models.Model):
-    test_plan_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255, unique=True)
-    creator_id = models.IntegerField()
-    milestone_id = models.IntegerField(null=True, blank=True)
-    description = models.TextField(null=True, blank=True)
-    test_case_filter = models.TextField(null=True, blank=True)
-
-class TestPlanFiles(models.Model):
-    file_id = models.AutoField(primary_key=True)
-    file_path = models.CharField(max_length=255)
-    test_plan = models.ForeignKey(TestPlan, on_delete=models.CASCADE)
-
-class TestPlanTicket(models.Model):
-    ticket_id = models.AutoField(primary_key=True)
-    ticket = models.CharField(max_length=255, unique=True)
-    test_plan = models.ForeignKey(TestPlan, on_delete=models.CASCADE)
-
-
-
-
-
-
-class User(models.Model):
-    user_id = models.AutoField(primary_key=True)
-    email = models.EmailField(unique=True)
-    first_name = models.CharField(max_length=64)
-    last_name = models.CharField(max_length=64)
-    profile_pic_path = models.CharField(max_length=256)
-
-class UserApiKey(models.Model):
-    key_id = models.AutoField(primary_key=True)
-    api_key = models.CharField(max_length=255, unique=True)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-class UserAccountIntegration(models.Model):
-    integration_id = models.AutoField(primary_key=True)
-    platform = models.CharField(max_length=255)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-
-
-
-
-
-class TestRunTestCase(models.Model):
-    test_run_test_case_id = models.AutoField(primary_key=True)
-    test_run = models.ForeignKey('TestRun', on_delete=models.CASCADE)
 
 
 
@@ -187,12 +103,6 @@ class TestRunTestCase(models.Model):
 
 
 
-class TestPlanTestCase(models.Model):
-    test_plan_test_case_id = models.AutoField(primary_key=True)
-    test_plan = models.ForeignKey('TestPlan', on_delete=models.CASCADE)
-    test_case = models.ForeignKey('TestCase', on_delete=models.CASCADE)
 
-class TestPlanTestSuite(models.Model):
-    test_plan_test_suite_id = models.AutoField(primary_key=True)
-    test_plan = models.ForeignKey('TestPlan', on_delete=models.CASCADE)
-    test_suite = models.ForeignKey('TestSuite', on_delete=models.CASCADE)
+
+
