+++ backend/drf_backend/tcms/serializers/test_case.py
@@ -72,91 +72,25 @@ class TestCaseResultsSerializer(serializers.ModelSerializer):
 
 
 
-class TestSuiteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestSuite
-        fields = '__all__'
-
-class SectionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Section
-        fields = '__all__'
-
-class TestSuiteFilesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestSuiteFiles
-        fields = '__all__'
-
-
-class TestRunSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestRun
-        fields = '__all__'
-
-class TestRunFilesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestRunFiles
-        fields = '__all__'
-
-class TestRunTicketSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestRunTicket
-        fields = '__all__'
-
-class TestPlanSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestPlan
-        fields = '__all__'
-
-class TestPlanFilesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestPlanFiles
-        fields = '__all__'
-
-class TestPlanTicketSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestPlanTicket
-        fields = '__all__'
 
 
 
 
 
 
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = '__all__'
-
-class UserApiKeySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserApiKey
-        fields = '__all__'
-
-class UserAccountIntegrationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserAccountIntegration
-        fields = '__all__'
 
 
 
 
 
-class TestRunTestCaseSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestRunTestCase
-        fields = '__all__'
 
 
 
 
 
-class TestPlanTestCaseSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestPlanTestCase
-        fields = '__all__'
 
-class TestPlanTestSuiteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = TestPlanTestSuite
-        fields = '__all__'
+
+
+
+
+
