+++ backend/drf_backend/tcms/views/test_case.py
@@ -55,79 +55,21 @@ class BugTrackerTicketsViewSet(viewsets.ModelViewSet):
 
 
 
-class TestSuiteViewSet(viewsets.ModelViewSet):
-    queryset = TestSuite.objects.all()
-    serializer_class = TestSuiteSerializer
-
-class SectionViewSet(viewsets.ModelViewSet):
-    queryset = Section.objects.all()
-    serializer_class = SectionSerializer
-
-class TestSuiteFilesViewSet(viewsets.ModelViewSet):
-    queryset = TestSuiteFiles.objects.all()
-    serializer_class = TestSuiteFilesSerializer
-
-
-
-
-
-class TestRunViewSet(viewsets.ModelViewSet):
-    queryset = TestRun.objects.all()
-    serializer_class = TestRunSerializer
-
-class TestRunFilesViewSet(viewsets.ModelViewSet):
-    queryset = TestRunFiles.objects.all()
-    serializer_class = TestRunFilesSerializer
-
-class TestRunTicketViewSet(viewsets.ModelViewSet):
-    queryset = TestRunTicket.objects.all()
-    serializer_class = TestRunTicketSerializer
-
-
-class TestPlanViewSet(viewsets.ModelViewSet):
-    queryset = TestPlan.objects.all()
-    serializer_class = TestPlanSerializer
-
-class TestPlanFilesViewSet(viewsets.ModelViewSet):
-    queryset = TestPlanFiles.objects.all()
-    serializer_class = TestPlanFilesSerializer
-
-class TestPlanTicketViewSet(viewsets.ModelViewSet):
-    queryset = TestPlanTicket.objects.all()
-    serializer_class = TestPlanTicketSerializer
-
-
-
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-class UserApiKeyViewSet(viewsets.ModelViewSet):
-    queryset = UserApiKey.objects.all()
-    serializer_class = UserApiKeySerializer
-
-class UserAccountIntegrationViewSet(viewsets.ModelViewSet):
-    queryset = UserAccountIntegration.objects.all()
-    serializer_class = UserAccountIntegrationSerializer
 
 
 
 
 
 
-class TestRunTestCaseViewSet(viewsets.ModelViewSet):
-    queryset = TestRunTestCase.objects.all()
-    serializer_class = TestRunTestCaseSerializer
 
 
 
 
-class TestPlanTestCaseViewSet(viewsets.ModelViewSet):
-    queryset = TestPlanTestCase.objects.all()
-    serializer_class = TestPlanTestCaseSerializer
 
-class TestPlanTestSuiteViewSet(viewsets.ModelViewSet):
-    queryset = TestPlanTestSuite.objects.all()
-    serializer_class = TestPlanTestSuiteSerializer
+
+
+
+
+
+
+
